Move `InvalidPaginationCursorError` to `dmr.pagination` and add this error to docs
