+++ dmr/exceptions.py
@@ -138,24 +138,3 @@ class TooManyRequestsError(Exception):
         super().__init__(msg or self.default_message)
         self.headers = headers
         self.error_type = ErrorType.ratelimit
-
-
-@final
-class InvalidPaginationCursorError(Exception):
-    """
-    Raised when the cursor passed for pagination is invalid.
-
-    See :class:`dmr.pagination.DjangoCursorPaginator._decode_cursor`
-    for more details.
-    """
-
-    default_message: ClassVar[str | Promise] = _('Invalid cursor')
-    status_code: ClassVar[HTTPStatus] = HTTPStatus.BAD_REQUEST
-
-    def __init__(self, msg: str | Promise | None = None) -> None:
-        """Provides default error message."""
-        # Circular import:
-        from dmr.errors import ErrorType  # noqa: PLC0415
-
-        super().__init__(msg or self.default_message)
-        self.error_type = ErrorType.value_error
